cnn-chip-flaw-detection/sept-scratch-model/requirements.txt
servicefoundry==0.6.1
requests==2.28.1
pandas==1.4.4
numpy==1.23.3
Pillow==9.2.0
opencv-python-headless==4.5.5.62
tensorflow==2.9.0
tensorflow-addons==0.17.1
